Organize Mongoose set up module.
