+++ console/src/app/modules/app-module/Routing/app-routing.module.ts
@@ -12,4 +12,3 @@ import { routes } from './routes';
   ]
 })
 export class AppRoutingModule { }
-
