+++ console/src/app/modules/mongoose-set-up-module/components/mongoose-set-up/mongoose-set-up.component.ts
@@ -1,11 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { MongooseSetupTab } from './mongoose-setup-tab.model';
+import { MongooseSetupTab } from '../../models/mongoose-setup-tab.model';
 // import { slideAnimation } from '../../../core/animations';
-import { RoutesList } from '../../app-module/Routing/routes-list';
+import { RoutesList } from '../../../app-module/Routing/routes-list';
 import { Subscription } from 'rxjs';
-import { Constants } from '../../../common/constants';
-import { MongooseSetUpService } from '../../../core/services/mongoose-set-up-service/mongoose-set-up.service';
+import { Constants } from '../../../../common/constants';
+import { MongooseSetUpService } from '../../../../core/services/mongoose-set-up-service/mongoose-set-up.service';
 
 @Component({
   selector: 'app-mongoose-set-up',
