+++ console/src/app/modules/mongoose-set-up-module/components/mongoose-set-up/set-up-steps/configuration-set-up/configuration-editing/configuration-editing.component.html
