+++ console/src/app/modules/mongoose-set-up-module/components/mongoose-set-up/set-up-steps/configuration-set-up/control-editing-root/control-editing-root.component.ts
