+++ console/src/app/modules/mongoose-set-up-module/components/mongoose-set-up/set-up-steps/scenarios-set-up/scenarios/scenarios.component.html
