+++ console/src/app/modules/mongoose-set-up-module/set-up-routing/mongoose-setup.routes.ts
@@ -1,9 +1,9 @@
 import { Routes } from "@angular/router";
 import { RoutesList } from "src/app/modules/app-module/Routing/routes-list";
-import { MongooseSetUpComponent } from "../mongoose-set-up/mongoose-set-up.component";
-import { NodesComponent } from "../mongoose-set-up/set-up-steps/nodes/nodes.component";
-import { ConfigurationEditingRootComponent } from "../mongoose-set-up/set-up-steps/configuration-set-up/control-editing-root/control-editing-root.component";
-import { ScenariosComponent } from "../mongoose-set-up/set-up-steps/scenarios-set-up/scenarios/scenarios.component";
+import { MongooseSetUpComponent } from "../components/mongoose-set-up/mongoose-set-up.component";
+import { NodesComponent } from "../components/mongoose-set-up/set-up-steps/nodes/nodes.component";
+import { ConfigurationEditingRootComponent } from "../components/mongoose-set-up/set-up-steps/configuration-set-up/control-editing-root/control-editing-root.component";
+import { ScenariosComponent } from "../components/mongoose-set-up/set-up-steps/scenarios-set-up/scenarios/scenarios.component";
 
 export const SETUP_ROUTES: Routes = [
   { path: RoutesList.NODES, component: NodesComponent },
